Testing

+++ Classificatore.h
@@ -21,7 +21,7 @@ class Classificatore
 public:
     virtual std::string Classify(Documento d)=0;
 //protected:
-    float Distanza(int tipo,Documento d1,Documento d2);
+    double Distanza(int tipo,Documento d1,Documento d2);
     std::vector<std::string> GetCommonKeys(std::map<std::string,int> m1,std::map<std::string,int> m2) const;
 };
